fix(Response): language available on code element, not pre (#16)
* fix(Response): language available on code element, not pre

Infer code language from the className on the code element (not from pre).

The className prop on the component is a string. The node.properties.classname is an array of class tokens. We use className with a regex to extract the language.

Default language plaintext. Happy to change this if JS is a better default.

* fix div inside pre

* fix(components): update regex for language extraction

Changed the regex in components.tsx to allow for non-whitespace characters in language class names, improving language detection for code blocks.

---------

Co-authored-by: Hayden Bleasel <hello@haydenbleasel.com>